Plot power spectrum over whole domain
Compute the power spectrum over the entire domain rather than at a
single point.

Each frame now comes with an array of times, to match the dimensions of
the x, y, u, v, w arrays.

A density plot is used to visualise the power spectrum using all of the
data.  Each x, z point gives a single 1d power spectrum. All of these
are plotted on the same chart, with a 2d histogram as the colour axis.
This gives an idea of the tendency of the power spectrum.

Finally, overlaid on the plot is a line of gradient -5/3 to represent
the Kolmogorov energy cascade.

Two useful topics on SO: [1] uses imshow to make a density plot. Here we
use pcolormesh as it is much more flexible; as in [2], we've used a
lognorm color scale (matplotlib.colors.LogNorm).

[1]: http://stackoverflow.com/questions/10439961/efficiently-create-a-density-plot-for-high-density-regions-points-for-sparse-re
[2]: http://stackoverflow.com/questions/11488800/non-linear-axes-for-imshow-in-matplotlib