fix non convergence of interpolator
Interpolation was failing on r13_12_16a.hdf5. This was a literal corner
case - in this run one of the array corners was nan. This leads to
errors as points are no longer guaranteed to be within the convex hull
of some interpolator.

First, we only write non-nan output from the interpolator to the data
array - this means only a single painting pass is needed.

Second, we apply nearest neighbour interpolation on the corners of the
array. We take a small slice at each corner and find the nearest non nan
neighbour. This isn't ideal, but we have to pin the interpolator
somewhere.

only copy non nan values over