split up long invalid slices
In some runs there is a contiguous region of nans that has a long time
extent. It takes a very long time to construct the interpolator for such
regions because of the number of points. Here we set a maximum time
extent for the slice size and split big slices up into a number of
smaller ones.