plot mean subtracted field, vertical mean
poss solution to wave subtraction: use the fact that the current is at
the bottom of the data field. the upper field is largely unaffected by
the current except in the mean. if we subtract the mean field from a
vertical slice and take a vertical mean of the upper ~10 rows of this
then we can get an indicator of wave activity external to the current.